Telligent 10 Compatibility
Update the date properties to use String rather than the built in
DateTime type.

There appears to be an issue in 10.1 where the date set in the UI isn't
read back properly due to the Javascript using an unrecognised date
format.

Update the properties to be simple strings in a standard date format and
update the plugin validation to handle invalid dates.